CPT-1149: improved variogram viz based on code review feedback and general preference for the direction of visualisation at this stage.
