+++ packages/evo-objects/src/evo/objects/typed/ellipsoid.py
@@ -9,7 +9,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-"""Ellipsoid and rotation primitives for spatial search operations."""
+"""Ellipsoid and rotation primitives for spatial operations."""
 
 from __future__ import annotations
 
@@ -29,7 +29,7 @@ __all__ = [
 
 
 def _rotation_matrix(dip_azimuth: float, dip: float, pitch: float) -> NDArray[np.floating[Any]]:
-    """Create a 3D rotation matrix from Leapfrog convention angles."""
+    """Create a 3D rotation matrix from Geoscience object convention angles."""
     az = np.radians(dip_azimuth)
     d = np.radians(dip)
     p = np.radians(pitch)
@@ -57,23 +57,11 @@ def _rotation_matrix(dip_azimuth: float, dip: float, pitch: float) -> NDArray[np
 
 @dataclass
 class EllipsoidRanges:
-    """The ranges (semi-axes lengths) of an ellipsoid.
-
-    Used to define the spatial extent of search neighborhoods in geostatistical
-    operations like kriging, simulation, and other estimation techniques.
-
-    Example:
-        >>> ranges = EllipsoidRanges(major=200.0, semi_major=150.0, minor=100.0)
-    """
+    """The ranges (semi-axes lengths) of an ellipsoid."""
 
     major: float
-    """The major axis length of the ellipsoid (largest extent)."""
-
     semi_major: float
-    """The semi-major axis length of the ellipsoid (intermediate extent)."""
-
     minor: float
-    """The minor axis length of the ellipsoid (smallest extent)."""
 
     def __init__(self, major: float, semi_major: float, minor: float):
         self.major = major
@@ -81,22 +69,9 @@ class EllipsoidRanges:
         self.minor = minor
 
     def to_dict(self) -> dict[str, Any]:
-        """Serialize to dictionary."""
-        return {
-            "major": self.major,
-            "semi_major": self.semi_major,
-            "minor": self.minor,
-        }
+        return {"major": self.major, "semi_major": self.semi_major, "minor": self.minor}
 
     def scaled(self, factor: float) -> "EllipsoidRanges":
-        """Return a new EllipsoidRanges scaled by the given factor.
-
-        Example:
-            >>> ranges = EllipsoidRanges(major=100, semi_major=50, minor=25)
-            >>> scaled = ranges.scaled(2.0)  # Double the size
-            >>> scaled.major
-            200.0
-        """
         return EllipsoidRanges(
             major=self.major * factor,
             semi_major=self.semi_major * factor,
@@ -106,25 +81,11 @@ class EllipsoidRanges:
 
 @dataclass
 class Rotation:
-    """The rotation of an ellipsoid using Leapfrog convention.
-
-    Defines the orientation of an ellipsoid in 3D space using three sequential
-    rotations: dip azimuth (about Z), dip (about X'), and pitch (about Z'').
-
-    Example:
-        >>> rotation = Rotation(dip_azimuth=45.0, dip=30.0, pitch=0.0)
-        >>> # Or use defaults (no rotation):
-        >>> rotation = Rotation()
-    """
+    """The rotation of an ellipsoid using Leapfrog/Geoscience Object convention."""
 
     dip_azimuth: float = 0.0
-    """First rotation, about the z-axis, in degrees (0-360)."""
-
     dip: float = 0.0
-    """Second rotation, about the x-axis, in degrees (0-90)."""
-
     pitch: float = 0.0
-    """Third rotation, about the z-axis, in degrees."""
 
     def __init__(self, dip_azimuth: float = 0.0, dip: float = 0.0, pitch: float = 0.0):
         self.dip_azimuth = dip_azimuth
@@ -132,60 +93,27 @@ class Rotation:
         self.pitch = pitch
 
     def to_dict(self) -> dict[str, Any]:
-        """Serialize to dictionary."""
-        return {
-            "dip_azimuth": self.dip_azimuth,
-            "dip": self.dip,
-            "pitch": self.pitch,
-        }
+        return {"dip_azimuth": self.dip_azimuth, "dip": self.dip, "pitch": self.pitch}
 
 
 @dataclass
 class Ellipsoid:
-    """An ellipsoid defining a spatial search region.
-
-    Combines ranges (semi-axes lengths) with rotation to define an oriented
-    ellipsoid in 3D space. Used for neighborhood searches in geostatistical
-    operations.
-
-    Example:
-        >>> ellipsoid = Ellipsoid(
-        ...     ranges=EllipsoidRanges(major=200.0, semi_major=150.0, minor=100.0),
-        ...     rotation=Rotation(dip_azimuth=45.0, dip=30.0, pitch=0.0),
-        ... )
-        >>>
-        >>> # Generate mesh for 3D visualization
-        >>> x, y, z = ellipsoid.surface_points(center=(100, 200, 50))
-    """
+    """An ellipsoid defining a spatial region."""
 
     ranges: EllipsoidRanges
-    """The ranges (semi-axes lengths) of the ellipsoid."""
-
     rotation: Rotation | None = None
-    """The rotation of the ellipsoid. Defaults to no rotation if not specified."""
 
     def __init__(self, ranges: EllipsoidRanges, rotation: Rotation | None = None):
         self.ranges = ranges
         self.rotation = rotation or Rotation()
 
     def to_dict(self) -> dict[str, Any]:
-        """Serialize to dictionary."""
         return {
             "ellipsoid_ranges": self.ranges.to_dict(),
             "rotation": self.rotation.to_dict() if self.rotation else Rotation().to_dict(),
         }
 
     def scaled(self, factor: float) -> "Ellipsoid":
-        """Return a new Ellipsoid scaled by the given factor.
-
-        The rotation is preserved, only the ranges are scaled.
-
-        Example:
-            >>> ell = Ellipsoid(EllipsoidRanges(100, 50, 25), Rotation(45, 30, 0))
-            >>> search_ell = ell.scaled(2.0)  # Create search ellipsoid 2x the variogram range
-            >>> search_ell.ranges.major
-            200.0
-        """
         return Ellipsoid(
             ranges=self.ranges.scaled(factor),
             rotation=Rotation(
@@ -200,24 +128,7 @@ class Ellipsoid:
         center: tuple[float, float, float] = (0, 0, 0),
         n_points: int = 20,
     ) -> tuple[NDArray[np.floating[Any]], NDArray[np.floating[Any]], NDArray[np.floating[Any]]]:
-        """Generate surface mesh points for 3D visualization.
-
-        Returns flattened x, y, z arrays suitable for Plotly Mesh3d or similar.
-
-        Args:
-            center: Center point (x, y, z) for the ellipsoid.
-            n_points: Number of points in each parametric direction.
-
-        Returns:
-            Tuple of (x, y, z) as 1D numpy arrays.
-
-        Example with Plotly:
-            >>> import plotly.graph_objects as go
-            >>> x, y, z = ellipsoid.surface_points(center=(100, 200, 50))
-            >>> mesh = go.Mesh3d(x=x, y=y, z=z, alphahull=0, opacity=0.3, color='blue')
-            >>> fig = go.Figure(data=[mesh])
-            >>> fig.show()
-        """
+        """Generate surface mesh points for 3D visualization."""
         rot = self.rotation or Rotation()
         rot_matrix = _rotation_matrix(rot.dip_azimuth, rot.dip, rot.pitch)
 
@@ -232,46 +143,23 @@ class Ellipsoid:
         points = np.array([x.flatten(), y.flatten(), z.flatten()])
         rotated = rot_matrix @ points
 
-        x_out = rotated[0] + center[0]
-        y_out = rotated[1] + center[1]
-        z_out = rotated[2] + center[2]
-
-        return x_out, y_out, z_out
+        return rotated[0] + center[0], rotated[1] + center[1], rotated[2] + center[2]
 
     def wireframe_points(
         self,
         center: tuple[float, float, float] = (0, 0, 0),
         n_points: int = 30,
     ) -> tuple[NDArray[np.floating[Any]], NDArray[np.floating[Any]], NDArray[np.floating[Any]]]:
-        """Generate wireframe points for 3D visualization.
-
-        Returns x, y, z arrays with NaN separators between line segments,
-        suitable for Plotly Scatter3d with mode='lines'.
-
-        Args:
-            center: Center point (x, y, z) for the ellipsoid.
-            n_points: Number of points per circle.
-
-        Returns:
-            Tuple of (x, y, z) as 1D numpy arrays with NaN separators.
-
-        Example with Plotly:
-            >>> import plotly.graph_objects as go
-            >>> x, y, z = ellipsoid.wireframe_points(center=(100, 200, 50))
-            >>> line = go.Scatter3d(x=x, y=y, z=z, mode='lines', line=dict(color='blue'))
-            >>> fig = go.Figure(data=[line])
-            >>> fig.show()
-        """
+        """Generate wireframe points for 3D visualization."""
         rot = self.rotation or Rotation()
         rot_matrix = _rotation_matrix(rot.dip_azimuth, rot.dip, rot.pitch)
-
         theta = np.linspace(0, 2 * np.pi, n_points)
 
         all_x: list[float] = []
         all_y: list[float] = []
         all_z: list[float] = []
 
-        # XY plane (major-semi_major)
+        # XY plane
         x = self.ranges.major * np.cos(theta)
         y = self.ranges.semi_major * np.sin(theta)
         z = np.zeros_like(theta)
@@ -281,7 +169,7 @@ class Ellipsoid:
         all_y.extend((rotated[1] + center[1]).tolist() + [np.nan])
         all_z.extend((rotated[2] + center[2]).tolist() + [np.nan])
 
-        # XZ plane (major-minor)
+        # XZ plane
         x = self.ranges.major * np.cos(theta)
         y = np.zeros_like(theta)
         z = self.ranges.minor * np.sin(theta)
@@ -291,7 +179,7 @@ class Ellipsoid:
         all_y.extend((rotated[1] + center[1]).tolist() + [np.nan])
         all_z.extend((rotated[2] + center[2]).tolist() + [np.nan])
 
-        # YZ plane (semi_major-minor)
+        # YZ plane
         x = np.zeros_like(theta)
         y = self.ranges.semi_major * np.cos(theta)
         z = self.ranges.minor * np.sin(theta)
@@ -302,3 +190,4 @@ class Ellipsoid:
         all_z.extend((rotated[2] + center[2]).tolist() + [np.nan])
 
         return np.array(all_x), np.array(all_y), np.array(all_z)
+
